feat(offline): create OfflineOverviewScreen
Adds a new `OfflineOverviewScreen` composable to inform the user that they are offline.

The screen displays a warning message and provides a button to view liked hunts, which are available for offline use. If the user chooses to view them, a list of `HuntCard`s is displayed. If no liked hunts are available, a corresponding message is shown.